# Faster onEachEvent event (including namespaces)