Roughly 30% faster with simple events:
```
new-observable#simple event x 1,801,581 ops/sec ±1.03% (87 runs sampled)
new-observable#namespaced event x 1,697,375 ops/sec ±4.13% (85 runs sampled)
new-observable#multiple events x 259,675 ops/sec ±1.62% (82 runs sampled)

old-observable#simple event x 1,385,113 ops/sec ±1.00% (88 runs sampled)
old-observable#multiple events x 238,766 ops/sec ±3.05% (81 runs sampled)
```